requirements.txt
Django==2.2.7
django-graphql-jwt==0.2.2
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.2.1
graphql-relay==0.4.5
iso8601==0.1.12
promise==2.2.1
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
